+++ deepreg/loss/image.py
@@ -1,9 +1,9 @@
 """Provide different loss or metrics classes for images."""
 import tensorflow as tf
 
-from deepreg.model.loss.util import NegativeLossMixin
-from deepreg.model.loss.util import gaussian_kernel1d_size as gaussian_kernel1d
-from deepreg.model.loss.util import (
+from deepreg.loss.util import NegativeLossMixin
+from deepreg.loss.util import gaussian_kernel1d_size as gaussian_kernel1d
+from deepreg.loss.util import (
     rectangular_kernel1d,
     separable_filter,
     triangular_kernel1d,
@@ -23,13 +23,13 @@ class SumSquaredDifference(tf.keras.losses.Loss):
 
     def __init__(
         self,
-        reduction: str = tf.keras.losses.Reduction.AUTO,
+        reduction: str = tf.keras.losses.Reduction.SUM,
         name: str = "SumSquaredDifference",
     ):
         """
         Init.
 
-        :param reduction: using AUTO reduction,
+        :param reduction: using SUM reduction over batch axis,
             calling the loss like `loss(y_true, y_pred)` will return a scalar tensor.
         :param name: name of the loss
         """
@@ -62,7 +62,7 @@ class GlobalMutualInformation(tf.keras.losses.Loss):
         self,
         num_bins: int = 23,
         sigma_ratio: float = 0.5,
-        reduction: str = tf.keras.losses.Reduction.AUTO,
+        reduction: str = tf.keras.losses.Reduction.SUM,
         name: str = "GlobalMutualInformation",
     ):
         """
@@ -70,7 +70,7 @@ class GlobalMutualInformation(tf.keras.losses.Loss):
 
         :param num_bins: number of bins for intensity, the default value is empirical.
         :param sigma_ratio: a hyper param for gaussian function
-        :param reduction: using AUTO reduction,
+        :param reduction: using SUM reduction over batch axis,
             calling the loss like `loss(y_true, y_pred)` will return a scalar tensor.
         :param name: name of the loss
         """
@@ -132,7 +132,7 @@ class GlobalMutualInformation(tf.keras.losses.Loss):
         div = (pab + EPS) / (papb + EPS)
         return tf.reduce_sum(pab * tf.math.log(div + EPS), axis=[1, 2])
 
-    def get_config(self):
+    def get_config(self) -> dict:
         """Return the config dictionary for recreating this class."""
         config = super().get_config()
         config["num_bins"] = self.num_bins
@@ -189,7 +189,7 @@ class LocalNormalizedCrossCorrelation(tf.keras.losses.Loss):
         self,
         kernel_size: int = 9,
         kernel_type: str = "rectangular",
-        reduction: str = tf.keras.losses.Reduction.AUTO,
+        reduction: str = tf.keras.losses.Reduction.SUM,
         name: str = "LocalNormalizedCrossCorrelation",
     ):
         """
@@ -197,7 +197,7 @@ class LocalNormalizedCrossCorrelation(tf.keras.losses.Loss):
 
         :param kernel_size: int. Kernel size or kernel sigma for kernel_type='gauss'.
         :param kernel_type: str, rectangular, triangular or gaussian
-        :param reduction: using AUTO reduction,
+        :param reduction: using SUM reduction over batch axis,
             calling the loss like `loss(y_true, y_pred)` will return a scalar tensor.
         :param name: name of the loss
         """
@@ -265,7 +265,7 @@ class LocalNormalizedCrossCorrelation(tf.keras.losses.Loss):
 
         return tf.reduce_mean(ncc, axis=[1, 2, 3, 4])
 
-    def get_config(self):
+    def get_config(self) -> dict:
         """Return the config dictionary for recreating this class."""
         config = super().get_config()
         config["kernel_size"] = self.kernel_size