+++ deepreg/model/backbone/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 from deepreg.model.backbone.global_net import GlobalNet
+from deepreg.model.backbone.interface import Backbone
 from deepreg.model.backbone.local_net import LocalNet
 from deepreg.model.backbone.u_net import UNet