+++ test/unit/test_loss_label.py
@@ -11,7 +11,7 @@ import numpy as np
 import pytest
 import tensorflow as tf
 
-import deepreg.model.loss.label as label
+import deepreg.loss.label as label
 
 
 class TestMultiScaleLoss:
@@ -26,7 +26,7 @@ class TestMultiScaleLoss:
         expected = dict(
             scales=None,
             kernel="gaussian",
-            reduction=tf.keras.losses.Reduction.AUTO,
+            reduction=tf.keras.losses.Reduction.SUM,
             name="MultiScaleLoss",
         )
         assert got == expected
@@ -71,7 +71,7 @@ class TestDiceScore:
             neg_weight=0.0,
             scales=None,
             kernel="gaussian",
-            reduction=tf.keras.losses.Reduction.AUTO,
+            reduction=tf.keras.losses.Reduction.SUM,
             name="DiceScore",
         )
         assert got == expected
@@ -112,7 +112,7 @@ class TestCrossEntropy:
             neg_weight=0.0,
             scales=None,
             kernel="gaussian",
-            reduction=tf.keras.losses.Reduction.AUTO,
+            reduction=tf.keras.losses.Reduction.SUM,
             name="CrossEntropy",
         )
         assert got == expected
@@ -155,7 +155,7 @@ class TestJaccardIndex:
             binary=False,
             scales=None,
             kernel="gaussian",
-            reduction=tf.keras.losses.Reduction.AUTO,
+            reduction=tf.keras.losses.Reduction.SUM,
             name="JaccardIndex",
         )
         assert got == expected